Annotate all backend commands for specta codegen
Complete the backend half of the typed-bindings migration (ADR-0009):
every #[tauri::command] now carries #[specta::specta] and every
Serialize-deriving struct that crosses the seam derives specta::Type. All
commands registered in the export-only builder; bindings.gen.ts now covers
102 commands and 35 types.

Two specta constraints handled at the source:
- create_annotation (14 params) exceeds tauri-specta SpectaFn arity. Carved
  out of the typed bindings; stays raw in generate_handler! and will be
  hand-wrapped in $lib/api.
- specta forbids BigInt-style types (usize/i64/...). Row-count returns
  changed usize -> u32; count fields kept at their Diesel type with
  #[specta(type = i32)] overrides. Values are all well within JS-safe range.

Runtime unchanged (generate_handler! still the source of truth). Frontend
call sites migrate next. 737 tests pass; svelte-check clean.

+++ src-tauri/src/commands/links.rs
@@ -338,7 +338,7 @@ pub fn get_alias_collisions(
     get_alias_collisions_on_conn(conn, note_id)
 }
 
-#[derive(Serialize, Debug, Clone)]
+#[derive(Serialize, specta::Type, Debug, Clone)]
 pub struct ResolvedNote {
     pub id: i32,
     pub title: String,
@@ -401,6 +401,7 @@ pub fn resolve_note_by_alias_on_conn(
 }
 
 #[tauri::command]
+#[specta::specta]
 pub fn resolve_note_by_alias(
     alias: String,
     ledger: State<AppLedger>,
@@ -539,7 +540,8 @@ struct BacklinkCountRow {
 }
 
 #[tauri::command]
-pub fn get_note_backlink_count(note_path: String, ledger: State<AppLedger>) -> Result<usize, String> {
+#[specta::specta]
+pub fn get_note_backlink_count(note_path: String, ledger: State<AppLedger>) -> Result<u32, String> {
     let mut state = ledger.lock().map_err(|_| "Ledger lock poisoned")?;
     let conn = state.connection.as_mut().ok_or("No ledger open")?;
     let row = diesel::sql_query(
@@ -548,7 +550,7 @@ pub fn get_note_backlink_count(note_path: String, ledger: State<AppLedger>) -> R
     .bind::<diesel::sql_types::Text, _>(&note_path)
     .load::<BacklinkCountRow>(conn)
     .map_err(|e| e.to_string())?;
-    Ok(row.into_iter().next().map(|r| r.cnt as usize).unwrap_or(0))
+    Ok(row.into_iter().next().map(|r| r.cnt as u32).unwrap_or(0))
 }
 
 #[tauri::command]
